fix(context): repair assembler budget split
